# Issue 4325: Sporadic OOMKilled in Segment Store under moderate workloads