--- a/segmentstore/storage/impl/src/main/java/io/pravega/segmentstore/storage/impl/rocksdb/RocksDBCache.java
+++ b/segmentstore/storage/impl/src/main/java/io/pravega/segmentstore/storage/impl/rocksdb/RocksDBCache.java
@@ -251,10 +251,7 @@ class RocksDBCache implements Cache {
         BlockBasedTableConfig tableFormatConfig = new BlockBasedTableConfig()
                 .setBlockSize(cacheBlockSizeKB * 1024L)
                 .setBlockCacheSize(readCacheSizeMB * 1024L * 1024L)
-                .setBlockCache(new LRUCache(readCacheSizeMB * 1024L * 1024L, 4))
-                .setCacheIndexAndFilterBlocks(true)
-                .setIndexType(IndexType.kHashSearch)
-                .setFilter(new BloomFilter());
+                .setCacheIndexAndFilterBlocks(true);
 
         Options options = new Options()
                 .setCreateIfMissing(true)
@@ -269,13 +266,10 @@ class RocksDBCache implements Cache {
                 .setOptimizeFiltersForHits(true)
                 .setUseDirectReads(this.directReads)
                 .setSkipStatsUpdateOnDbOpen(true)
-                .optimizeForPointLookup(readCacheSizeMB * 1024L * 1024L)
                 .setIncreaseParallelism(INTERNAL_ROCKSDB_PARALLELISM)
                 .setMaxBackgroundFlushes(INTERNAL_ROCKSDB_PARALLELISM / 2)
                 .setMaxBackgroundJobs(INTERNAL_ROCKSDB_PARALLELISM)
-                .setCompactionStyle(CompactionStyle.LEVEL)
-                .setMaxBackgroundCompactions(INTERNAL_ROCKSDB_PARALLELISM)
-                .setLevelCompactionDynamicLevelBytes(true);
+                .setMaxBackgroundCompactions(INTERNAL_ROCKSDB_PARALLELISM);
 
         if (this.memoryOnly) {
             Env env = new RocksMemEnv();