**Change log description**  
Partial revert of new RocksDB config (leave only new parallelism settings).

**Purpose of the change**  
Fixes #4325.

**What the code does**  
In PR #4173, we changed several RocksDB settings to prevent reads from the cache to be significantly slower than writes, in the presence of unthrottled writers. The main change to solve this issue was to increase the internal parallelism of RocksDB (e.g., compactions jobs), but other changes were also introduced based on the suggestions of the RocksDB performance config guide (https://github.com/facebook/rocksdb/wiki/RocksDB-Tuning-Guide). Unfortunately, the execution of multiple benchmarks with Pravega 0.6 (RocksDB on disk) demonstrated that these extra changes induce instability in the Segment Store (`OOMKilled` in pods) under moderate/high workloads.

This PR reverts most of the changes from PR #4173 and leaves only the parallelism-related configuration changes, which were the most relevant ones to solve issue #4158.
 
**How to verify it**  
We verified that adding only parallelism-related changes makes no impact on the memory usage of the Segment Store (e.g., no `OOMKilled`) on AWS EKS, while we retain the performance improvements induced by PR #4173. In the following, we compare the performance of Pravega before #4173 with the changes of this PR:

```
Experiment: 100B events / throttled 5MBps / 5 benchmark pods / 5 writers/pod
Before #4173: avg=92.10ms, p50=75.00ms, p75=113.00ms, p95=230.00, p99=312.00, e/s=261885.055, MBps=25.00
This PR:      avg=40.00ms, p50=32.00ms, p75=46.00ms, p95=81.00, p99=143.00, e/s=242617.95, MBps=24.15
``` 
```
Experiment: 1000B events / throttled 40MBps / 5 benchmark pods / 5 writers/pod
Before #4173: avg=171.8ms, p50=123ms, p75=194ms, p95=410, p99=876, e/s=202,801.38, MBps=193.4
This PR:      avg=73.20ms, p50=55ms, p75=87ms, p95=194, p99=325, e/s=209,674.34, MBps=199.95
``` 

```
Experiment: 10000B events / throttled 64MBps / 5 benchmark pods / 5 writers/pod
Before #4173: avg=235.7ms, p50=75ms, p75=179ms, p95=1446, p99=1958, e/s=33,477.37, MBps=319.25
This PR:      avg=37ms, p50=11ms, p75=31ms, p95=166, p99=335, e/s=33,548.49, MBps=319.95
``` 

```
Experiment: 100000B events / throttled 88MBps / 5 benchmark pods / 5 writers/pod
Before #4173: avg=166.4ms, p50=33ms, p75=146ms, p95=820, p99=1312, e/s=4523.25, MBps=431.35
This PR:      avg=46.9ms, p50=14ms, p75=52ms, p95=182, p99=314, e/s=4594.785, MBps=438.2
``` 

Moreover, during the execution of all these (and more) benchmarks, we have not seen any Segment Store pod restart:
```
λ kubectl.exe get pods
NAME                                               READY   STATUS    RESTARTS   AGE
benchmark-pod-0                                    1/1     Running   0          3d2h
benchmark-pod-1                                    1/1     Running   0          3d2h
benchmark-pod-2                                    1/1     Running   0          3d2h
benchmark-pod-3                                    1/1     Running   0          3d2h
benchmark-pod-4                                    1/1     Running   0          3d2h
fio                                                1/1     Running   0          21d
grafana-84f6fb6588-mn6g2                           1/1     Running   0          25d
influxdb-868dd94b5d-skcfp                          1/1     Running   0          26d
local-volume-provisioner-5dkqp                     1/1     Running   0          20d
local-volume-provisioner-b4bx9                     1/1     Running   0          20d
local-volume-provisioner-kmb8k                     1/1     Running   0          20d
local-volume-provisioner-mcp9w                     1/1     Running   0          20d
local-volume-provisioner-s2tnt                     1/1     Running   0          20d
pravega-bookie-0                                   1/1     Running   1          150m
pravega-bookie-1                                   1/1     Running   0          150m
pravega-bookie-2                                   1/1     Running   1          150m
pravega-operator-5cd44fc7c5-sjs4m                  1/1     Running   0          5d2h
pravega-pravega-controller-67cc4fb6f6-5xjmv        1/1     Running   0          150m
pravega-pravega-segmentstore-0                     1/1     Running   0          150m
pravega-pravega-segmentstore-1                     1/1     Running   0          149m
pravega-pravega-segmentstore-2                     1/1     Running   0          149m
volted-tortoise-efs-provisioner-56cb5c4f75-7nb8c   1/1     Running   3          26d
zookeeper-0                                        1/1     Running   0          152m
zookeeper-1                                        1/1     Running   0          151m
zookeeper-2                                        1/1     Running   0          151m
zookeeper-operator-5bb86b8d59-qxjnh                1/1     Running   0          5d2h
```

All unit/integration tests should be passing as before.

Signed-off-by: Raúl Gracia <raul.gracia@emc.com>